refactor(suite-native): simplify receive navigation structure
Refs #29733
