+++ suite-native/add-coin-account/src/hooks/useAddCoinAccount.ts
@@ -208,7 +208,7 @@ export const useAddCoinAccount = (networksSearchQuery?: string) => {
                 break;
             case 'home':
                 navigation.replace(RootStackRoutes.ReceiveStack, {
-                    screen: ReceiveStackRoutes.ReceiveAccount,
+                    screen: ReceiveStackRoutes.ReceiveAddress,
                     params: {
                         networkSymbol: symbol,
                         accountType,
@@ -227,7 +227,7 @@ export const useAddCoinAccount = (networksSearchQuery?: string) => {
                 break;
             case 'receive':
                 navigation.navigate(RootStackRoutes.ReceiveStack, {
-                    screen: ReceiveStackRoutes.ReceiveAccount,
+                    screen: ReceiveStackRoutes.ReceiveAddress,
                     params: {
                         networkSymbol: symbol,
                         accountType,
