+++ suite-native/module-receive/src/screens/ReceiveAddressScreen.tsx
@@ -3,15 +3,26 @@ import { FadeOut } from 'react-native-reanimated';
 import { useSelector } from 'react-redux';
 
 import { G } from '@mobily/ts-belt';
-import { useNavigation, usePreventRemove } from '@react-navigation/native';
+import {
+    type RouteProp,
+    useNavigation,
+    usePreventRemove,
+    useRoute,
+} from '@react-navigation/native';
 
-import { selectIsDeviceBackupRequired } from '@suite-common/device';
+import { type DeviceRootState, selectIsDeviceBackupRequired } from '@suite-common/device';
 import { getDisplaySymbol } from '@suite-common/wallet-config';
-import { type AccountsRootState, selectAccountByKey } from '@suite-common/wallet-core';
+import {
+    type AccountsRootState,
+    selectAccountByKey,
+    selectDeviceAccountKeyForNetworkSymbolAndAccountTypeWithIndex,
+} from '@suite-common/wallet-core';
 import { type AccountKey, type TokenAddress } from '@suite-common/wallet-types';
 import { AccountDetailsCard } from '@suite-native/accounts';
 import {
     AnimatedBox,
+    BoxSkeleton,
+    Card,
     ErrorMessage,
     InlineAlertBox,
     VStack,
@@ -26,9 +37,16 @@ import { useInAppRating } from '@suite-native/in-app-rating';
 import { Translation } from '@suite-native/intl';
 import { Link } from '@suite-native/link';
 import { WalletBackupNotSetWarningBottomSheet } from '@suite-native/module-device-onboarding';
-import { type CloseActionType } from '@suite-native/navigation';
+import {
+    type CloseActionType,
+    type ReceiveStackParamList,
+    type ReceiveStackRoutes,
+    Screen,
+} from '@suite-native/navigation';
 import { type TokensRootState, selectAccountTokenSymbol } from '@suite-native/tokens';
 import TrezorConnect from '@trezor/connect';
+import { getScreenWidth } from '@trezor/env-utils';
+import { prepareNativeStyle, useNativeStyles } from '@trezor/styles-native';
 import { HELP_CENTER_OTHER_CRYPTOCURRENCIES_DESTINATION_TAGS_URL } from '@trezor/urls';
 
 import { ReceiveBlockedDeviceCompromisedScreen } from './ReceiveBlockedDeviceCompromisedScreen';
@@ -37,17 +55,53 @@ import { ReceiveScreenHeader } from '../components/ReceiveScreenHeader';
 import { hasReceiveAddressButtonRequest } from '../hooks/receiveSelectors';
 import { useAccountReceiveAddress } from '../hooks/useAccountReceiveAddress';
 
-type ReceiveAddressScreenProps = {
+const SCREEN_WIDTH = getScreenWidth();
+
+const cardStyle = prepareNativeStyle(utils => ({
+    padding: utils.spacings.sp8,
+}));
+
+type ReceiveAddressContentProps = {
     accountKey: AccountKey;
     tokenContract?: TokenAddress;
     closeActionType: CloseActionType;
 };
 
-export const ReceiveAddressScreen = ({
+const ReceiveAddressLoader = ({
+    tokenContract,
+    closeActionType,
+}: Pick<ReceiveAddressContentProps, 'tokenContract' | 'closeActionType'>) => {
+    const { applyStyle } = useNativeStyles();
+
+    return (
+        <Screen
+            header={
+                <ReceiveScreenHeader
+                    tokenContract={tokenContract}
+                    closeActionType={closeActionType}
+                />
+            }
+        >
+            <VStack spacing="sp32" alignItems="center" paddingHorizontal="sp8">
+                <Card style={applyStyle(cardStyle)}>
+                    <BoxSkeleton width={SCREEN_WIDTH - 32} height={70} />
+                </Card>
+                <Card style={applyStyle(cardStyle)}>
+                    <VStack spacing="sp24" alignItems="center" paddingHorizontal="sp24">
+                        <BoxSkeleton width={SCREEN_WIDTH - 80} height={200} />
+                        <BoxSkeleton width={SCREEN_WIDTH - 80} height={48} borderRadius={24} />
+                    </VStack>
+                </Card>
+            </VStack>
+        </Screen>
+    );
+};
+
+const ReceiveAddressContent = ({
     accountKey,
     tokenContract,
     closeActionType,
-}: ReceiveAddressScreenProps) => {
+}: ReceiveAddressContentProps) => {
     const { askForRating } = useInAppRating();
     const account = useSelector((state: AccountsRootState) =>
         selectAccountByKey(state, accountKey),
@@ -186,5 +240,43 @@ export const ReceiveAddressScreen = ({
                 />
             )}
         </ConfirmOnTrezorWrapper>
+    );
+};
+
+export const ReceiveAddressScreen = () => {
+    const {
+        params: {
+            accountKey: routeAccountKey,
+            tokenContract,
+            networkSymbol: routeNetworkSymbol,
+            accountType: routeAccountType,
+            accountIndex: routeAccountIndex,
+            closeActionType,
+        },
+    } = useRoute<RouteProp<ReceiveStackParamList, ReceiveStackRoutes.ReceiveAddress>>();
+
+    const foundAccountKey = useSelector((state: AccountsRootState & DeviceRootState) =>
+        selectDeviceAccountKeyForNetworkSymbolAndAccountTypeWithIndex(
+            state,
+            routeNetworkSymbol,
+            routeAccountType,
+            routeAccountIndex,
+        ),
+    );
+
+    const accountKey = routeAccountKey ?? foundAccountKey;
+
+    if (!accountKey) {
+        return (
+            <ReceiveAddressLoader tokenContract={tokenContract} closeActionType={closeActionType} />
+        );
+    }
+
+    return (
+        <ReceiveAddressContent
+            accountKey={accountKey}
+            tokenContract={tokenContract}
+            closeActionType={closeActionType}
+        />
     );
 };
