+++ suite-native/navigation/src/routes.ts
@@ -261,7 +261,7 @@ export enum YieldStackRoutes {
 
 export enum ReceiveStackRoutes {
     ReceiveAccounts = 'ReceiveAccounts',
-    ReceiveAccount = 'ReceiveAccount',
+    ReceiveAddress = 'ReceiveAddress',
 }
 
 export enum SendStackRoutes {
